requirements.txt
azure-cognitiveservices-speech
keyboard
pyautogui
pycaw
python-dotenv
